# Modified the script to call terraform show less